+++ react-tagsinput.js
@@ -62,6 +62,8 @@
       , onChange: React.PropTypes.func
       , onChangeInput: React.PropTypes.func
       , onBlur: React.PropTypes.func
+      , onKeyDown: React.PropTypes.func
+      , onKeyUp: React.PropTypes.func
       , onTagAdd: React.PropTypes.func
       , onTagRemove: React.PropTypes.func
       , transform: React.PropTypes.func
@@ -80,6 +82,8 @@
         , onChange: function () { }
         , onChangeInput: function () { }
         , onBlur: function () { }
+        , onKeyDown: function () { }
+        , onKeyUp: function () { }
         , onTagAdd: function () { }
         , onTagRemove: function () { }
         , transform: function (tag) { return tag.trim(); }
@@ -222,6 +226,10 @@
       if (remove && this._value().length > 0 && this.state.tag === "") {
         this.removeTag(this._value()[this._value().length - 1]);
       }
+
+      if (!add && !remove) {
+        this.props.onKeyDown(e);
+      }
     }
 
     , onChange: function (e) {
@@ -244,8 +252,18 @@
       this.props.onBlur();
     }
 
+    , clear: function () {
+      this.setState({ tag: "", invalid: false });
+    }
+
     , focus: function () {
       this.refs.input.getDOMNode().focus();
+    }
+
+    , handleClick: function (e) {
+      if (e.target === this.getDOMNode()) {
+        this.focus();
+      }
     }
 
     , render: function() {
@@ -263,6 +281,7 @@
       return (
         React.createElement("div", {
           className: ns + "tagsinput"
+          , onClick: this.handleClick
         }, tagNodes, React.createElement(Input, {
           ref: "input"
           , ns: ns
@@ -271,6 +290,7 @@
           , invalid: this.state.invalid
           , validating: this.state.validating
           , onKeyDown: this.onKeyDown
+          , onKeyUp: this.props.onKeyUp
           , onChange: this.onChange
           , onBlur: this.onBlur
         }))